+++ packages/ui/src/components/library/HTable/HTable.vue
@@ -1,7 +1,7 @@
 <template>
   <q-table
     :loading="loading"
-    :separator="settings.display.table.separator"
+    :separator="separator"
     :dense="settings.display.table.dense"
     :rows-per-page-options="rowsPerPageOptions"
     :rows="rows"
@@ -14,6 +14,7 @@
 
     <template v-if="!$slots['top-right']" #top-right="props">
       <h-table-action
+        v-model="separator"
         :inFullscreen="props.inFullscreen"
         @toggle-fullscreen="props.toggleFullscreen"
       ></h-table-action>
@@ -45,11 +46,11 @@
   </q-table>
 </template>
 
-<script lang="ts">
-import type { PropType } from 'vue';
-import { defineComponent, computed } from 'vue';
+<script setup lang="ts">
+import type { PropType, ShallowRef } from 'vue';
+import { computed, shallowRef } from 'vue';
 
-import type { Entity } from '@/lib/declarations';
+import type { Entity, QTableSeparatorProps } from '@/lib/declarations';
 
 import { useSettingsStore } from '@/stores';
 import { useDictionary } from '@/composables/constants';
@@ -58,49 +59,39 @@ import HTableAction from './HTableAction.vue';
 import HStatusColumn from './HStatusColumn.vue';
 import HReservedColumn from './HReservedColumn.vue';
 
-export default defineComponent({
+defineOptions({
   name: 'HTable',
-
   components: {
     HReservedColumn,
     HStatusColumn,
     HTableAction,
   },
+});
 
-  emits: ['update:pageNumber'],
+const props = defineProps({
+  rows: { type: Array as PropType<Array<Entity>>, required: true },
+  pageNumber: { type: Number, default: 0 },
+  totalPages: { type: Number },
+  loading: { type: Boolean, default: false },
+  showAll: { type: Boolean, default: false },
+  status: { type: Boolean, default: false },
+  reserved: { type: Boolean, default: false },
+});
 
-  props: {
-    rows: { type: Array as PropType<Array<Entity>>, required: true },
-    pageNumber: { type: Number, default: 0 },
-    totalPages: { type: Number },
-    loading: { type: Boolean, default: false },
-    showAll: { type: Boolean, default: false },
-    status: { type: Boolean, default: false },
-    reserved: { type: Boolean, default: false },
+const emit = defineEmits(['update:pageNumber']);
+
+const settings = useSettingsStore();
+const { options } = useDictionary('DataItemStatus');
+const separator = shallowRef('horizontal') as ShallowRef<QTableSeparatorProps>;
+
+const pageNumberVModel = computed({
+  get: () => props.pageNumber,
+  set: (newValue) => {
+    emit('update:pageNumber', newValue);
   },
+});
 
-  setup(props, { emit }) {
-    const settings = useSettingsStore();
-
-    const { options } = useDictionary('DataItemStatus');
-
-    const pageNumberVModel = computed({
-      get: () => props.pageNumber,
-      set: (newValue) => {
-        emit('update:pageNumber', newValue);
-      },
-    });
-
-    const rowsPerPageOptions = computed(() => {
-      return props.showAll ? [0] : [5, 10, 15, 20, 25, 50];
-    });
-
-    return {
-      settings,
-      pageNumberVModel,
-      rowsPerPageOptions,
-      options,
-    };
-  },
+const rowsPerPageOptions = computed(() => {
+  return props.showAll ? [0] : [5, 10, 15, 20, 25, 50];
 });
 </script>