+++ packages/ui/src/lib/declarations/index.ts
@@ -2,5 +2,4 @@ export * from './base';
 export * from './components';
 export * from './route';
 export * from './customize';
-export * from './settings';
 export * from './message';