+++ packages/ui/src/lib/definitions/enums.ts
@@ -1,25 +1,6 @@
-export { ContentTypeEnum } from '@herodotus-cloud/core';
-export { SocialSourceEnum, CaptchaCategoryEnum, NotificationCategoryEnum } from '@herodotus-cloud/apis';
-
-export enum OperationEnum {
-  CREATE = 'create',
-  EDIT = 'edit',
-  AUTHORIZE = 'authorize',
-  INFO = 'info',
-  ALLOCATABLE = 'allocatable',
-  SETUP = 'setup',
-  INVOKE = 'invoke',
-}
-
-export enum LayoutModeEnum {
-  DEFAULT = 'defaults',
-  CLASSIC = 'classic',
-  TRANSVERSE = 'transverse',
-  COLUMNS = 'transverse',
-}
-
-export enum ThemeModeEnum {
-  DARK = 'dark',
-  LIGHT = 'light',
-  SYSTEM = 'system',
-}
+export { ContentTypeEnum, OperationEnum } from '@herodotus-cloud/core';
+export {
+  SocialSourceEnum,
+  CaptchaCategoryEnum,
+  NotificationCategoryEnum,
+} from '@herodotus-cloud/apis';