+++ shared/framework-kernel/src/lib/hooks/index.ts
@@ -1,0 +1,4 @@
+import useEditFinish from './useEditFinish';
+import useSystemTheme from './useSystemTheme';
+
+export { useEditFinish, useSystemTheme };